+++ packages/core/src/scripts/utils/create-app.ts
@@ -1,24 +1,34 @@
-import open from "open";
 import { MultiCompiler } from "webpack";
 import express from "express";
 import createServer from "./create-server";
-import { Mode } from "../../../types";
+import { openBrowserTab } from "./open-browser";
+import { DevOptions } from "../dev";
 
-// Create an express app ready to be used with webpack-dev-middleware.
+/**
+ * Create an express app ready to be used with webpack-dev-middleware.
+ *
+ * @param options - Defined in {@link DevOptions}.
+ *
+ * @returns - An object with the app and a callback function that starts the
+ * server when Webpack has finished.
+ */
 export default async ({
   mode,
   port,
   isHttps,
   target,
   openBrowser = true,
-}: {
-  mode: Mode;
-  port: number;
-  isHttps: boolean;
-  target: "es5" | "module";
-  openBrowser?: boolean;
-}): Promise<{
+  publicPath = "/static",
+}: DevOptions): Promise<{
+  /**
+   * The app created by Express, ready to be used by webpack-dev-middleware.
+   */
   app: express.Express;
+
+  /**
+   * A function that accepts a {@link MultiCompiler} and starts the server
+   * once Webpack has finished the bundles.
+   */
   done: (compiler: MultiCompiler) => void;
 }> => {
   // Create the app.
@@ -50,14 +60,19 @@ export default async ({
   // Start listening.
   server.listen(port, () => {
     console.log(
-      `\n\nSERVER STARTED -- Listening @ ${url}\n  - mode: ${mode}\n  - target: ${target}\n\n`
+      `\n\nSERVER STARTED -- Listening @ ${url}\n  - mode: ${mode}\n  - target: ${target}\n  - public-path: ${publicPath}\n\n`
     );
   });
 
   // Open localhost on the local browser.
-  if (openBrowser) open(url);
+  if (openBrowser) openBrowserTab(url);
 
-  // Check if webpack has finished (both the client and server bundles).
+  //
+  /**
+   * Check if webpack has finished (both the client and server bundles).
+   *
+   * @param compiler - A {@link MultiCompiler} from Webpack.
+   */
   const done = (compiler: MultiCompiler) => {
     compiler.compilers[0].hooks.done.tapAsync(
       "frontity-dev-server",