+++ packages/frontity/src/cli/index.ts
@@ -1,35 +1,14 @@
 #! /usr/bin/env node
-import * as tsNode from "ts-node";
 
-tsNode.register({
-  transpileOnly: true,
-  compilerOptions: {
-    // Target latest version of ECMAScript.
-    target: "es2017",
-    // Search under node_modules for non-relative imports.
-    moduleResolution: "node",
-    // commonjs modules.
-    module: "commonjs",
-    // Allow default imports from modules with no default export.
-    allowSyntheticDefaultImports: true,
-    // Don't emit; allow Babel to transform files.
-    noEmit: true,
-    // Import non-ES modules as default imports.
-    esModuleInterop: true,
-    // Resolve JSON files.
-    resolveJsonModule: true,
-    // Support for JSX.
-    jsx: "react",
-    // Transpile JS as well.
-    allowJs: true,
-  },
-});
+require("dotenv").config();
 
 import program from "commander";
-import { dev, build, serve } from "../commands";
 
 import { default as create } from "./create";
 import { default as createPackage } from "./create-package";
+import { default as dev } from "./dev";
+import { default as build } from "./build";
+import { default as serve } from "./serve";
 import { default as subscribe } from "./subscribe";
 import { default as unknown } from "./unknown";
 import { default as info } from "./info";
@@ -47,25 +26,25 @@ program
 // options: --typescript, --use-cwd.
 program
   .command("create [name]")
-  .option("-h, --theme <theme>", "The theme to use")
-  .option("-t, --typescript", "Adds support for TypeScript")
-  .option("-c, --use-cwd", "Generates the project in the current directory.")
-  .option("-n, --no-prompt", "Skips prompting the user for options")
+  .option("--theme <theme>", "The theme to use")
+  .option("--typescript", "Adds support for TypeScript")
+  .option("--use-cwd", "Generates the project in the current directory.")
+  .option("--no-prompt", "Skips prompting the user for options")
   .description("Creates a new Frontity project.")
   .action((name, { ...args }) => create({ name, ...args }));
 
 program
   .command("create-package [name]")
-  .option("-n, --namespace <value>", "Sets the namespace for this package")
+  .option("--namespace <value>", "Sets the namespace for this package")
   .option("--no-prompt", "Skips prompting the user for options")
   .description("Creates a new Frontity package in a project.")
   .action((name, { ...args }) => createPackage({ name, ...args }));
 
 program
   .command("dev")
-  .option("-p, --production", "Builds the project for production.")
+  .option("--prod, --production", "Builds the project for production.")
   .option("--port <port>", "Runs the server on a custom port. Default is 3000.")
-  .option("-s, --https", "Runs the server using https.")
+  .option("--https", "Runs the server using https.")
   .option(
     "--dont-open-browser",
     "Don't open a browser window with the localhost."
@@ -74,9 +53,9 @@ program
     "--target <target>",
     'create bundles with "es5" or "module". Default target is "module".'
   )
-
+  .option("--publicPath <path>", "DEPRECATED, use --public-path instead.")
   .option(
-    "--publicPath <path>",
+    "--public-path <path>",
     'set the public path for static assets. Default path is "/static/".'
   )
   .description("Starts a server in development mode.")
@@ -84,13 +63,14 @@ program
 
 program
   .command("build")
-  .option("-d, --development", "Builds the project for development.")
+  .option("--dev, --development", "Builds the project for development.")
   .option(
     "--target <target>",
-    'create bundles with "es5" or "module". Default target is "module".'
+    'create bundles with "es5", "module" or "both". Default target is "both".'
   )
+  .option("--publicPath <path>", "DEPRECATED, use --public-path instead.")
   .option(
-    "--publicPath <path>",
+    "--public-path <path>",
     'set the public path for static assets. Default path is "/static/".'
   )
   .description("Builds the project for production.")
@@ -99,7 +79,7 @@ program
 program
   .command("serve")
   .option("--port <port>", "Runs the server on a custom port. Default is 3000.")
-  .option("-s, --https", "Runs the server using https.")
+  .option("--https", "Runs the server using https.")
   .description("Starts a server in production mode.")
   .action(serve);