+++ packages/wp-source/src/libraries/schemas/__tests__/list.tests.ts
@@ -1,14 +1,17 @@
 import { normalize } from "normalizr";
-// https://demo.worona.org/?rest_route=/wp/v2/posts&categories=7&_embed=true
+// https://test.frontity.org/?rest_route=/wp/v2/posts&categories=7&_embed=true
 import category7 from "./mocks/posts-from-category-7.json";
-// https://demo.worona.org/?rest_route=/wp/v2/categories&_embed=true&per_page=3
+// https://test.frontity.org/?rest_route=/wp/v2/categories&_embed=true&per_page=3
 import categoriesList from "./mocks/categories-list.json";
-// https://demo.worona.org/?rest_route=/wp/v2/tags&_embed=true
+// https://test.frontity.org/?rest_route=/wp/v2/tags&_embed=true
 import tagsList from "./mocks/tags-list.json";
-// https://demo.worona.org/?rest_route=/wp/v2/users&_embed=true
+// https://test.frontity.org/?rest_route=/wp/v2/users&_embed=true
 import authorList from "./mocks/author-list.json";
-// https://demo.worona.org/?rest_route=/wp/v2/media&_embed=true
+// https://test.frontity.org/?rest_route=/wp/v2/media&_embed=true
 import mediaList from "./mocks/media-list.json";
+// https://test.frontity.org/?rest_route=/wp/v2/media&_embed=true
+import commentList from "./mocks/post-60-comments.json";
+
 import { list } from "../";
 
 test("Convert list of posts using list", () => {
@@ -33,5 +36,10 @@ test("Convert a author list", () => {
 
 test("Convert a media list", () => {
   const { entities } = normalize(mediaList, list);
+  expect(entities).toMatchSnapshot();
+});
+
+test("Convert a comment list", () => {
+  const { entities } = normalize(commentList, list);
   expect(entities).toMatchSnapshot();
 });