+++ packages/html2react/CHANGELOG.md
@@ -1,5 +1,27 @@
 # Change Log
 
+## 1.4.0
+
+### Minor Changes
+
+- [`b12a1006`](https://github.com/frontity/frontity/commit/b12a100643d3207ab540b66cd7a1bb5626953900) [#548](https://github.com/frontity/frontity/pull/548) Thanks [@DAreRodz](https://github.com/DAreRodz)! - Add a `processors` prop to the `Html2React` component which overrides the processors defined in `libraries.html2react.processors` for that specific element.
+
+## 1.3.5
+
+### Patch Changes
+
+- [`c062b39c`](https://github.com/frontity/frontity/commit/c062b39c9091ba40edc56e00a8c92cdd542ad9a5) [#530](https://github.com/frontity/frontity/pull/530) Thanks [@orballo](https://github.com/orballo)! - Fixes html2react parser support for some SVG attributes that contain dashes and colons
+
+## 1.3.4
+
+### Patch Changes
+
+- [`19912053`](https://github.com/frontity/frontity/commit/1991205379a103bc4ec09397cca99c1f2ba09a2d) [#486](https://github.com/frontity/frontity/pull/486) Thanks [@juanmaguitar](https://github.com/juanmaguitar)! - README.md added.
+
+- Updated dependencies [[`cfa8062d`](https://github.com/frontity/frontity/commit/cfa8062d60ccf46b3e0c5aa5490e4e03f1b693a5), [`ba13f70a`](https://github.com/frontity/frontity/commit/ba13f70ae2a4360ca21c77aed1c920c02e9d45b8), [`845de536`](https://github.com/frontity/frontity/commit/845de536d111105dd3f636c3543170a676ac65cc), [`62fce1e5`](https://github.com/frontity/frontity/commit/62fce1e5c117faeb5902dc0ddae3b13d95cd925b)]:
+  - @frontity/components@1.3.3
+  - frontity@1.9.0
+
 ## 1.3.3
 
 ### Patch Changes